feat: fetch用のクラスとして分離
