+++ modules/data/src/main/java/net/pantasystem/milktea/data/di/module/NodeInfoModule.kt
@@ -6,6 +6,8 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import net.pantasystem.milktea.data.api.NodeInfoAPIBuilder
 import net.pantasystem.milktea.data.api.NodeInfoAPIBuilderImpl
+import net.pantasystem.milktea.data.infrastructure.nodeinfo.NodeInfoFetcher
+import net.pantasystem.milktea.data.infrastructure.nodeinfo.NodeInfoFetcherImpl
 import net.pantasystem.milktea.data.infrastructure.nodeinfo.NodeInfoRepositoryImpl
 import net.pantasystem.milktea.model.nodeinfo.NodeInfoRepository
 import javax.inject.Singleton
@@ -21,5 +23,9 @@ abstract class NodeInfoModule {
     @Binds
     @Singleton
     abstract fun nodeInfoAPIBuilder(impl: NodeInfoAPIBuilderImpl): NodeInfoAPIBuilder
+
+    @Binds
+    @Singleton
+    abstract fun bindNodeInfoFetcher(impl: NodeInfoFetcherImpl): NodeInfoFetcher
 
 }
